Break up SaveModal into smaller components
