+++ src/core/components/modals/Save/index.ts
@@ -1,0 +1,3 @@
+import SaveModal from './Save'
+
+export default SaveModal;
